use 'get_host_ip' on more places
